pyproject.toml
[project]
name = "real_python"
version = "0.1.0"
description = ""
authors = [
    {name = "leblancr",email = "rich.leblanc1@proton.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
